Refactor trajectory parsing into modular loaders
- split monolithic trajectory parsing into format-specific modules (ASE, HDF5, LAMMPS, VASP, XYZ)
- add unified frame reader, format detection, and shared helper utilities for indexed loading
- preserve explicit frame step values and validate sampled indexed frame metadata correctly
